SB-92 (Refactor) : GoodsService abandonment 가독성 Refactoring
SB-92 (Refactor) : GoodsService abandonment 가독성 Refactoring

goodsRepository.findAllByReceivingIdOrderByIdDesc(receivingId).forEach() -> 변수 저장 후 Empty 확인
